[WIP] Base for ts and a public archive page
Here's the base of my ideas for a public archive page and typescript.

It's definitely a mess, and the routes and db stuff around the public
archive will need some work. I was thinking collections, but it's
probably easier just to add a field to the archive schema. That might
also lead to us changing how we handle the open vs archive as well, but
that's a later discussion.

I set it up as a new collection so I wouldn't have to write migrations
yet. That's something I still need to model out and set up. I might
right it as a separate library, so don't worry about it, and we can get
that handled manually (it shouldn't be too bad, and I can help).

As for the typescript, I've really just included a tsconfig from another
project I worked on. We can totally rework that.

I'd prefer to split this commit or drop it later, but I almost forgot to
make this at all, so you're getting everything, sorry.